+++ lizard_area/urls.py
@@ -5,8 +5,10 @@ from django.conf.urls.defaults import url
 from django.contrib import admin
 
 from lizard_ui.urls import debugmode_urlpatterns
-from lizard_area.views import ApiView
-from lizard_area.views import Homepage
+from lizard_area.views import (
+    Homepage,
+    ApiView,
+)
 
 admin.autodiscover()
 
@@ -18,9 +20,11 @@ urlpatterns = patterns(
     url(r'^$', Homepage.as_view(), name=NAME_PREFIX + 'homepage'),
     url(r'^api-view/$', ApiView.as_view(), name=NAME_PREFIX + 'api_view'),
     (r'^api/', include('lizard_area.api.urls')),
-    (r'^area_links_panel/$',
-    'lizard_area.views.area_links_panel',
-     {},
-     "lizard_area.area_links_panel"),
-    )
+    (
+        r'^area_links_panel/$',
+        'lizard_area.views.area_links_panel',
+        {},
+        "lizard_area.area_links_panel"
+    ),
+)
 urlpatterns += debugmode_urlpatterns()